Add PLCTag abstraction layer over the C library
